Fix: Clean up lift mdp, unify the soft and core lift mdp.
