--- source/isaaclab_tasks/isaaclab_tasks/core/lift/config/franka_soft/mdp/__init__.py
@@ -1,10 +1,0 @@
-# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-"""This sub-module contains the functions that are specific to the deformable lift environments."""
-
-from isaaclab.utils.module import lazy_export
-
-lazy_export()
