--- source/isaaclab_tasks/isaaclab_tasks/core/lift/config/franka_soft/mdp/__init__.pyi
@@ -1,46 +1,0 @@
-# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-__all__ = [
-    # rewards
-    "deformable_lifted",
-    "deformable_lifting",
-    "deformable_ee_distance",
-    "deformable_com_ee_distance",
-    "deformable_fingertip_distance",
-    "deformable_com_goal_distance",
-    "deformable_com_goal_reached",
-    # terminations
-    "deformable_outside_bounds",
-    "joint_vel_out_of_sim_limit",
-    # events
-    "reset_deformable_over_support",
-    # commands
-    "DeformableUniformPoseCommand",
-    "DeformableUniformPoseCommandCfg",
-    # curriculums
-    "gravity_range_linear",
-]
-
-from .curriculums import gravity_range_linear
-from .events import reset_deformable_over_support
-from .pose_commands import (
-    DeformableUniformPoseCommand,
-    DeformableUniformPoseCommandCfg,
-)
-from .rewards import (
-    deformable_com_ee_distance,
-    deformable_com_goal_distance,
-    deformable_com_goal_reached,
-    deformable_ee_distance,
-    deformable_fingertip_distance,
-    deformable_lifted,
-    deformable_lifting,
-)
-from .terminations import (
-    deformable_outside_bounds,
-    joint_vel_out_of_sim_limit,
-)
-from isaaclab_tasks.core.lift.mdp import *
