+++ source/isaaclab_tasks/isaaclab_tasks/core/lift/mdp/commands/pose_commands.py
@@ -13,15 +13,16 @@ from typing import TYPE_CHECKING
 
 import torch
 
+from isaaclab.assets import AssetBaseCfg
 from isaaclab.managers import CommandTerm
 from isaaclab.utils.leapp import POSE7_ELEMENT_NAMES
 from isaaclab.utils.math import combine_frame_transforms, compute_pose_error, quat_from_euler_xyz, quat_unique
 
 if TYPE_CHECKING:
-    from isaaclab.assets import Articulation, RigidObject
+    from isaaclab.assets import Articulation, DeformableObject, RigidObject
     from isaaclab.envs import ManagerBasedEnv
 
-    from . import pose_commands_cfg as dex_cmd_cfgs
+    from .pose_commands_cfg import DeformableUniformPoseCommandCfg, ObjectUniformPoseCommandCfg
 
 
 class ObjectUniformPoseCommand(CommandTerm):
@@ -49,10 +50,10 @@ class ObjectUniformPoseCommand(CommandTerm):
         and optional visualization settings.
     """
 
-    cfg: dex_cmd_cfgs.ObjectUniformPoseCommandCfg
+    cfg: ObjectUniformPoseCommandCfg
     """Configuration for the command generator."""
 
-    def __init__(self, cfg: dex_cmd_cfgs.ObjectUniformPoseCommandCfg, env: ManagerBasedEnv):
+    def __init__(self, cfg: ObjectUniformPoseCommandCfg, env: ManagerBasedEnv):
         """Initialize the command generator class.
 
         Args:
@@ -194,3 +195,55 @@ class ObjectUniformPoseCommand(CommandTerm):
             self.goal_visualizer.visualize(self.pose_command_w[:, :3], marker_indices=success_id + 1)
             # -- current object position
             self.curr_visualizer.visualize(obj_pos, marker_indices=success_id + 1)
+
+
+class DeformableUniformPoseCommand(ObjectUniformPoseCommand):
+    """Uniform position command for a deformable object, tracked by its center of mass.
+
+    Deformable objects expose no root orientation, so the target is tracked with the COM
+    (:attr:`~isaaclab.assets.DeformableObject.data.root_pos_w`) and only ``position_only``
+    commands are supported.
+
+    The success visualizer asset may be a static asset (``AssetBaseCfg``), which has no
+    runtime view. In that case its world position is the fixed spawn offset from the
+    environment origins.
+    """
+
+    cfg: DeformableUniformPoseCommandCfg
+    """Configuration for the command generator."""
+
+    object: DeformableObject
+    """The deformable object tracked by the command."""
+
+    def __init__(self, cfg: DeformableUniformPoseCommandCfg, env: ManagerBasedEnv):
+        if not cfg.position_only:
+            raise ValueError("DeformableUniformPoseCommand only supports position_only commands.")
+        super().__init__(cfg, env)
+
+        # static assets are stored as their config, so their world position is constant
+        if isinstance(self.success_vis_asset, AssetBaseCfg):
+            offset = torch.tensor(self.success_vis_asset.init_state.pos, device=self.device)
+            self._static_success_vis_pos_w = env.scene.env_origins + offset
+        else:
+            self._static_success_vis_pos_w = None
+
+    def _update_metrics(self):
+        # transform command from base frame to simulation world frame
+        self.pose_command_w[:, :3], self.pose_command_w[:, 3:] = combine_frame_transforms(
+            self.robot.data.root_pos_w.torch,
+            self.robot.data.root_quat_w.torch,
+            self.pose_command_b[:, :3],
+            self.pose_command_b[:, 3:],
+        )
+        com_w = self.object.data.root_pos_w.torch
+        self.metrics["position_error"] = torch.linalg.norm(self.pose_command_w[:, :3] - com_w, dim=-1)
+
+        if self.success_vis_asset is None:
+            return
+        # same success radius as the goal markers of the base class
+        success_id = (self.metrics["position_error"] < 0.05).int()
+        if self._static_success_vis_pos_w is not None:
+            vis_pos_w = self._static_success_vis_pos_w
+        else:
+            vis_pos_w = self.success_vis_asset.data.root_pos_w.torch
+        self.success_visualizer.visualize(vis_pos_w, marker_indices=success_id)
